# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Star Mall service.",
        "title": "Star Mall API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/products/create": {
            "put": {
                "description": "创建商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create 创建商品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ProductCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ProductCreateResponse"
                        }
                    },
                    "401": {
                        "description": "invalid token claims",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/products/getProduct/:id": {
            "get": {
                "description": "获取商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "GetProduct 获取商品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "商品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product get successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    },
                    "401": {
                        "description": "invalid token claims",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/products/search": {
            "post": {
                "description": "搜索商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "SearchProduct 搜索商品",
                "parameters": [
                    {
                        "description": "product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SearchProductRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product updated successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Product"
                            }
                        }
                    },
                    "401": {
                        "description": "没有权限",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/products/search/:MerchantID": {
            "get": {
                "description": "通过商家获取商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "GetProductByMerchant 通过商家获取商品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ProductCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product get successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.Product"
                        }
                    },
                    "401": {
                        "description": "invalid token claims",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/products/update": {
            "patch": {
                "description": "更新商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update 更新商品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ProductUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "没有权限",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Product": {
            "type": "object",
            "properties": {
                "attr": {
                    "description": "商品属性",
                    "type": "string"
                },
                "bookingTime": {
                    "description": "预售时间",
                    "type": "integer"
                },
                "brand": {
                    "description": "品牌",
                    "type": "string"
                },
                "cateID": {
                    "description": "小类ID",
                    "type": "integer"
                },
                "clickCount": {
                    "description": "点击量",
                    "type": "integer"
                },
                "content": {
                    "description": "商品内容",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "deletedAt": {
                    "description": "删除时间",
                    "type": "integer"
                },
                "desc": {
                    "description": "商品描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isBest": {
                    "description": "是否精品",
                    "type": "boolean"
                },
                "isBooking": {
                    "description": "是否预售",
                    "type": "boolean"
                },
                "isDeleted": {
                    "description": "是否删除",
                    "type": "integer"
                },
                "isHot": {
                    "description": "是否热门",
                    "type": "integer"
                },
                "isNew": {
                    "description": "是否新品",
                    "type": "boolean"
                },
                "keywords": {
                    "description": "商品关键词",
                    "type": "string"
                },
                "marketPrice": {
                    "description": "市场价",
                    "type": "number"
                },
                "merchantID": {
                    "description": "商家ID",
                    "type": "integer"
                },
                "price": {
                    "description": "价格",
                    "type": "number"
                },
                "productNum": {
                    "description": "商品数量",
                    "type": "integer"
                },
                "productSn": {
                    "description": "商品编号",
                    "type": "string"
                },
                "productTypeID": {
                    "description": "商品大类ID",
                    "type": "integer"
                },
                "purchaseCount": {
                    "description": "购买量",
                    "type": "integer"
                },
                "sort": {
                    "description": "排序权重",
                    "type": "integer"
                },
                "specs": {
                    "description": "商品规格",
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "subTitle": {
                    "type": "string"
                },
                "title": {
                    "description": "ImageID       string  `db:\"image_id\"`",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "integer"
                },
                "version": {
                    "description": "商品版本",
                    "type": "string"
                }
            }
        },
...
        "handler.ProductCreateRequest": {
            "type": "object",
            "properties": {
                "attr": {
                    "description": "@Description 商品属性",
                    "type": "string"
                },
                "bookingTime": {
                    "description": "@Description 预售时间",
                    "type": "integer"
                },
                "brand": {
                    "description": "@Description 品牌",
                    "type": "string"
                },
                "cateID": {
                    "description": "@Description 商品分类ID",
                    "type": "integer"
                },
                "content": {
                    "description": "@Description 商品内容",
                    "type": "string"
                },
                "desc": {
                    "description": "@Description 商品描述",
                    "type": "string"
                },
                "image_id": {
                    "description": "@Description 商品图片id数组",
                    "type": "string"
                },
                "isBooking": {
                    "description": "@Description 是否预售",
                    "type": "boolean"
                },
                "keywords": {
                    "description": "@Description 关键字",
                    "type": "string"
                },
                "marketPrice": {
                    "description": "@Description 商品市场价",
                    "type": "number"
                },
                "price": {
                    "description": "@Description 商品价格",
                    "type": "number"
                },
                "productTypeID": {
                    "description": "@Description 商品大类ID",
                    "type": "integer"
                },
                "sn": {
                    "description": "@Description 商品编号",
                    "type": "string"
                },
                "specs": {
                    "description": "@Description 商品规格",
                    "type": "string"
                },
                "subTitle": {
                    "description": "@Description 商品副标题",
                    "type": "string"
                },
                "title": {
                    "description": "@Description 商品名称",
                    "type": "string"
                },
                "version": {
                    "description": "@Description 商品版本",
                    "type": "string"
                }
            }
        },
        "handler.ProductCreateResponse": {
            "description": "创建商品响应体",
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 状态码",
                    "type": "integer"
                },
                "id": {
                    "description": "@Description 商品ID",
                    "type": "integer"
                },
                "message": {
                    "description": "@Description 消息",
                    "type": "string"
                }
            }
        },
        "handler.ProductUpdateRequest": {
            "description": "更新商品请求体",
            "type": "object",
            "properties": {
                "attr": {
                    "description": "@Description 属性",
                    "type": "string"
                },
                "booking_time": {
                    "description": "@Description 预售时间",
                    "type": "integer"
                },
                "cate_id": {
                    "description": "@Description 分类ID",
                    "type": "integer"
                },
                "content": {
                    "description": "@Description 内容",
                    "type": "string"
                },
                "desc": {
                    "description": "@Description 描述",
                    "type": "string"
                },
                "id": {
                    "description": "@Description 商品ID",
                    "type": "integer"
                },
                "isBooking": {
                    "description": "@Description 是否预售",
                    "type": "boolean"
                },
                "keywords": {
                    "description": "@Description 关键字",
                    "type": "string"
                },
                "market_price": {
                    "description": "@Description 市场价",
                    "type": "number"
                },
                "price": {
                    "description": "@Description 价格",
                    "type": "number"
                },
                "product_sn": {
                    "description": "@Description 商品编号",
                    "type": "string"
                },
                "product_type_id": {
                    "description": "@Description 商品类型ID",
                    "type": "integer"
                },
                "sub_title": {
                    "description": "@Description 子标题",
                    "type": "string"
                },
                "title": {
                    "description": "@Description 标题",
                    "type": "string"
                }
            }
        },
        "handler.SearchProductRequest": {
            "description": "搜索商品请求体",
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "输入 JWT token，格式为：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}